refactor: move hash routing onload function location
